MW-232 Remove duplicated code in forms
