+++ src/pages/Register/RegisterForm.tsx
@@ -1,19 +1,13 @@
 import { HowToReg } from '@mui/icons-material';
-import { Skeleton } from '@mui/material';
 import { useRegister } from '../../api/user';
-import FormikSwitch from '../../components/formikFields/FormikSwitch';
-import FormikTextField from '../../components/formikFields/FormikTextField';
-import { ALLOWED_IMAGE_FORMATS, ALLOWED_IMAGE_OBJECT } from '../../const';
-import { registerValidationSchema } from '../../data/formData/user';
+import {
+  RegisterFormValues,
+  UserFormFields,
+  registerValidationSchema,
+} from '../../data/formData/user';
 import { AccountType, PostUserDetails } from '../../types/UserTypes';
 import { getErrorMessage, toBase64 } from '../../utils/utils';
 import BaseForm from '../Login/BaseForm';
-import FormikFileUploader from './../../components/formikFields/FormikFileUploader';
-
-type RegisterFormValues = Pick<
-  PostUserDetails,
-  'email' | 'nickname' | 'name' | 'surname' | 'userType' | 'password'
-> & { avatarImage: File | null; repeatPassword?: string };
 
 const formikInitialValues = {
   email: '',
@@ -25,31 +19,6 @@ const formikInitialValues = {
   userType: AccountType.Simple,
   avatarImage: null,
 };
-
-const formFields = (
-  <>
-    <FormikTextField name='email' label='E-mail' type='email' />
-    <FormikTextField name='nickname' label='Nazwa użytkownika' />
-    <FormikTextField name='name' label='Imię' />
-    <FormikTextField name='surname' label='Nazwisko' />
-    <FormikTextField name='password' label='Hasło' type='password' />
-    <FormikTextField name='repeatPassword' label='Powtórz hasło' type='password' />
-    <FormikSwitch
-      name='userType'
-      labels={['Widz', 'Twórca']}
-      options={[AccountType.Simple, AccountType.Creator]}
-    />
-    <FormikFileUploader
-      name='avatarImage'
-      label='Avatar'
-      acceptedFileTypes={ALLOWED_IMAGE_FORMATS}
-      acceptObject={ALLOWED_IMAGE_OBJECT}
-      preview
-      previewProps={{ sx: { height: 70, width: 70 } }}
-      previewSkeleton={<Skeleton variant='circular' sx={{ width: 70, height: 70 }} />}
-    />
-  </>
-);
 
 function RegisterForm() {
   const { isLoading, error, mutate } = useRegister();
@@ -67,7 +36,7 @@ function RegisterForm() {
       title='Rejestracja'
       buttonText='Zarejestruj się'
       icon={<HowToReg />}
-      formFields={formFields}
+      formFields={<UserFormFields showRegisterFields />}
       initialValues={formikInitialValues}
       validationSchema={registerValidationSchema}
       onSubmit={handleSubmit}
