+++ src/pages/Upload/VideoUploadForm.tsx
@@ -1,18 +1,8 @@
 import { Publish } from '@mui/icons-material';
-import { Skeleton } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
 import { useVideoUpload } from '../../api/video';
-import FormikAutocomplete from '../../components/formikFields/FormikAutocomplete';
-import FormikFileUploader from '../../components/formikFields/FormikFileUploader';
-import FormikSwitch from '../../components/formikFields/FormikSwitch';
-import FormikTextField from '../../components/formikFields/FormikTextField';
 import {
-  ALLOWED_IMAGE_FORMATS,
-  ALLOWED_IMAGE_OBJECT,
-  ALLOWED_VIDEO_FORMATS,
-  ALLOWED_VIDEO_OBJECT,
-} from '../../const';
-import {
+  MetadataFormFields,
   MetadataFormValues,
   videoUploadValidationSchema,
 } from '../../data/formData/video';
@@ -38,61 +28,6 @@ function VideoUploadForm() {
   const navigate = useNavigate();
   const { mutate, error, isLoading } = useVideoUpload();
 
-  const formFields = (
-    <>
-      <FormikTextField name='title' label='Tytuł' />
-      <FormikTextField name='description' label='Opis' />
-      <FormikAutocomplete
-        name='tags'
-        multiple
-        freeSolo
-        autoSelect
-        optionsPromiseFn={() =>
-          import('../../const/predefined_tags').then((module) => module.default)
-        }
-        loadingText='Pobieranie sugestii...'
-        onInputChange={(event, value) => {
-          if (value.endsWith(',')) {
-            const target = event.currentTarget as HTMLInputElement;
-            target.blur();
-            target.focus();
-          }
-        }}
-        TextFieldProps={{
-          label: 'Tagi',
-        }}
-      />
-      <FormikFileUploader
-        name='thumbnail'
-        label='Miniaturka'
-        acceptedFileTypes={ALLOWED_IMAGE_FORMATS}
-        acceptObject={ALLOWED_IMAGE_OBJECT}
-        preview
-        previewProps={{
-          sx: {
-            height: 70,
-            width: 124,
-          },
-          variant: 'rounded',
-        }}
-        previewSkeleton={
-          <Skeleton variant='rounded' sx={{ aspectRatio: '16 / 9', height: 70 }} />
-        }
-      />
-      <FormikFileUploader
-        name='videoFile'
-        label='Wideo'
-        acceptedFileTypes={ALLOWED_VIDEO_FORMATS}
-        acceptObject={ALLOWED_VIDEO_OBJECT}
-      />
-      <FormikSwitch
-        name='visibility'
-        labels={['Prywatny', 'Publiczny']}
-        options={[VideoVisibility.Private, VideoVisibility.Public]}
-      />
-    </>
-  );
-
   const handleSubmit = async (values: VideoUploadFormValues) => {
     const formData = new FormData();
     formData.append('videoFile', values.videoFile as File, values.videoFile?.name);
@@ -110,7 +45,7 @@ function VideoUploadForm() {
       title='Publikacja wideo'
       buttonText='Publikuj'
       icon={<Publish />}
-      formFields={formFields}
+      formFields={<MetadataFormFields showVideoField />}
       initialValues={formikInitialValues}
       validationSchema={videoUploadValidationSchema}
       onSubmit={handleSubmit}
